+++ frontend/src/components/overview/Series.jsx
@@ -3,7 +3,7 @@ import React, { useEffect } from 'react';
 import { useParams } from 'react-router';
 import { useStateValue } from '../../contexts/state';
 import TimeLineChart from '../graphs/TimeLineChart';
-import DashboardList from '../dashlist/ListMain';
+import DashboardList from '../stabilityTable/ListMain';
 import LastBuildElement from '../lastBuildTable/LastBuildTable';
 import { useTranslation } from 'react-i18next';
 import {
