+++ frontend/src/components/stabilityTable/FailureTable.jsx
