+++ frontend/src/components/stabilityTable/FailureTable.styles.js
