+++ frontend/src/components/stabilityTable/ListMain.jsx
