+++ frontend/src/components/stabilityTable/ListMain.styles.js
